feat: add python compatibility
