+++ examples/web/react-tlock-demo/config-overrides.d.ts
