+++ examples/web/react-tlock-demo/public/favicon.ico
