+++ examples/web/react-tlock-demo/public/logo192.png
