+++ examples/web/react-tlock-demo/public/logo512.png
