+++ examples/web/react-tlock-demo/public/manifest.json
