+++ examples/web/react-tlock-demo/src/setupTests.d.ts
