+++ py/pyproject.toml
@@ -1,0 +1,26 @@
+[project]
+name = "timelock"
+version = "0.0.1"
+authors = [
+  { name="Ideal Labs", email="hello@idealabs.network" },
+]
+description = "This provides python bindings for usage of timelock encryption"
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
+[tool.setuptools]
+packages = ["timelock"]
+package-dir = { ""="src" }
+
+[project.urls]
+Homepage = "https://github.com/ideal-lab5/timelock"
+Issues = "https://github.com/ideal-lab5/timelock/issues"
+
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
