+++ py/src/timelock/__init__.py
@@ -1,0 +1,1 @@
+from .timelock import Timelock
