+++ wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "timelock-wasm-wrapper"
+name = "timelock_wasm_wrapper"
 version = "0.0.1"
 edition = "2021"
 license = "Apache-2.0"
@@ -37,6 +37,8 @@ rand_chacha = { version = "0.3.1", default-features = false }
 w3f-bls = { version = "0.1.3", default-features = false }
 sp-consensus-beefy-etf = { git = "https://github.com/ideal-lab5/idn-sdk.git", features = ["bls-experimental"]}
 sha2 = { version = "0.10.2", default-features = false }
+pyo3 = { version = "0.23.1", features = ["extension-module", "macros"], default-features = false }
+libc = "0.2"  # Make sure to use the latest version available
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.0"
