+++ wasm/README.md
@@ -1,12 +1,25 @@
-# WASM compat
+# WASM Bindings for Timelock
 
-This crate provides wasm compatibility for the timelock crate.
+This crate provides wasm compatibility for the timelock crate. It provides compatibility for both javascript and python (3+).
 
 ## Build
 
+### For JavaScript Compatibility
 To compile to wasm, first build the project and them run wasm-pack
 
 ``` shell
 cargo build
 wasm-pack build --target web --out-dir pkg
+```
+
+### For Python Compatibility
+
+Python compatible wasm output is generated with [PyO3](https://pyo3.rs/v0.23.2/).
+
+First create a virtual env, then run:
+``` sh
+pip install maturin
+# specify your python version
+export PYO3_CROSS_PYTHON_VERSION="3.10"
+maturin develop
 ```
