+++ wasm/wasm_build.sh
@@ -1,3 +1,3 @@
 cargo build
 # todo: create issue to investigate bundler instead
-wasm-pack build --target web --out-dir pkg
+wasm-pack build --target web --out-dir pkg/js
